requirements.txt
pytest>=7.0
pytest-asyncio>=0.21
grpcio>=1.50
grpcio-tools>=1.50
kuksa-client>=0.4
paho-mqtt>=1.6
pyyaml>=6.0
